vr-3d-object-detector/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
torch==2.5.1
torchvision==0.20.1
ultralytics==8.1.0
pillow==10.2.0
numpy==1.26.3
python-multipart==0.0.6
